Implement all generate command handlers to set the custom tool
