+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateAutoRestCommandHandler.cs
@@ -1,10 +1,10 @@
-﻿using System.IO;
-using Cairo;
-using MonoDevelop.Core;
+﻿using ApiClientCodeGen.VSMac.CustomTools.AutoRest;
 
 namespace ApiClientCodeGen.VSMac.Commands.Handlers
 {
     public class GenerateAutoRestCommandHandler : GenerateCommandHandler
     {
+        protected override string GeneratorName
+            => AutoRestSingleFileCustomTool.GeneratorName;
     }
 }
