+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateCommandHandler.cs
@@ -1,3 +1,4 @@
+using System;
 using MonoDevelop.Components.Commands;
 using MonoDevelop.Core;
 using MonoDevelop.Ide;
@@ -7,8 +8,12 @@ namespace ApiClientCodeGen.VSMac.Commands.Handlers
 {
     public abstract class GenerateCommandHandler : BaseCommandHandler
     {
+        protected abstract string GeneratorName { get; }
         protected virtual string SupportedFileExtension => ".json";
         protected FilePath FilePath { get; private set; }
+
+        protected override void Run() => SetGenerator();
+        protected override void Run(object dataItem) => SetGenerator();
         
         protected override void Update(CommandInfo info)
         {
@@ -16,8 +21,15 @@ namespace ApiClientCodeGen.VSMac.Commands.Handlers
             if (item == null)
                 return;
 
-            info.Visible = item.Name.EndsWith(SupportedFileExtension, System.StringComparison.OrdinalIgnoreCase) == true;
+            info.Visible = item.Name.EndsWith(SupportedFileExtension, StringComparison.OrdinalIgnoreCase);
             FilePath = item.FilePath;
+        }
+
+        private void SetGenerator()
+        {
+            var project = IdeApp.ProjectOperations.CurrentSelectedProject;
+            var item = project.Files.GetFile(FilePath);
+            item.Generator = GeneratorName;
         }
     }
 }
