+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateNSwagCommandHandler.cs
@@ -1,19 +1,10 @@
 using ApiClientCodeGen.VSMac.CustomTools.NSwag;
-using MonoDevelop.Ide;
 
 namespace ApiClientCodeGen.VSMac.Commands.Handlers
 {
     public class GenerateNSwagCommandHandler : GenerateCommandHandler
     {
-        protected override void Run() => SetGenerator();
-
-        protected override void Run(object dataItem) => SetGenerator();
-
-        private void SetGenerator()
-        {
-            var project = IdeApp.ProjectOperations.CurrentSelectedProject;
-            var item = project.Files.GetFile(FilePath);
-            item.Generator = NSwagSingleFileCustomTool.GeneratorName;
-        }
+        protected override string GeneratorName
+            => NSwagSingleFileCustomTool.GeneratorName;
     }
 }
