+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateOpenApiCommandHandler.cs
@@ -1,6 +1,10 @@
+using ApiClientCodeGen.VSMac.CustomTools.OpenApi;
+
 namespace ApiClientCodeGen.VSMac.Commands.Handlers
 {
     public class GenerateOpenApiCommandHandler : GenerateCommandHandler
     {
+        protected override string GeneratorName
+            => OpenApiSingleFileCustomTool.GeneratorName;
     }
 }
