+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateSwaggerCommandHandler.cs
@@ -1,6 +1,10 @@
+using ApiClientCodeGen.VSMac.CustomTools.Swagger;
+
 namespace ApiClientCodeGen.VSMac.Commands.Handlers
 {
     public class GenerateSwaggerCommandHandler : GenerateCommandHandler
     {
+        protected override string GeneratorName 
+            => SwaggerSingleFileCustomTool.GeneratorName;
     }
 }
